src/dcubelearning_tp/requirements.txt
azure-functions
azure-identity==1.12.0
azure-keyvault-secrets==4.7.0
